Implemented positional encoding mechanism
The PositionalEncoding class will be used by the Transformer to encode
its input using a continous trigonometric function. This will introduce
a new component in the input values, that will be correlated with each
value's position in an input sequence. Using this information, the model
will be able to learn positional relations between words.