Implemented input mask generation and finalized Transormer class
The input mask is a special tensor added to transformer's input in order
to zero-out tokens in front of the current token. This prevents the
model from calculating attention with "future" words. The input mask
needs to be a square 2D tensor with its dimension equal to the length of
a single input sequence.

Other thing to note is that the TransformerEncoder implementation
requires input to be in the form:

[sequence_len, batch_len, embedding_dim]

which is different from how the batches were created in the original
Udacity project. This will require either changing the batching code or
transposing the inputs.