First training attempt of the transformer
After completing the training functions, the transformer model was
constructed and trained. The process was slower compared to RNN training
and the loss was greater. These results are not what was expected. Also,
the dialog generation performed by the (somewhat) trained model yielded
very poor results.

In conclusion, it seems that either the hyperparameters were way off the
mark, or there is a flaw in the model and/or training pipeline.